460 Dependency Resolver With GraphQL:
Cleanup and testing file repository.
