+++ Package/Image/ImageResolution.cs
@@ -1,0 +1,15 @@
+using System.Linq;
+
+namespace OpenTap.Package
+{
+    class ImageResolution
+    {
+        public ImageResolution(PackageSpecifier[] pkgs, long iterations)
+        {
+            Packages = pkgs.OrderBy(x => x.Name).ToArray();
+            this.Iterations = iterations;
+        }
+        public readonly PackageSpecifier[] Packages;
+        public readonly long Iterations;
+    }
+}
